Use connection microservice instance in login_helper
The login_helper function from Login microservice used to call functions specified in default_controller.py instead of using the microservice instance directly. This meant that the Login microservice needed to be changed every time the server architecture changed. This commit makes Login independent of the server architecture, as long as the structure of bus.microservices remains constant.